Fix bug with some disallowed nodes
This fixes a scenario when a disallowed node does not have children. Current code will evaluate `node.children` and add undefined to  `parent.children`'s array. This should not happen because the consumer code relies on a child not being `undefined`.

This problem manifests only when mode is `unwrap`.

An example of child with this problem is `thematicBreak` (`source: "***"`)