Rework based on Unified, TypeScript, and Storybook
Unified, through remark and rehype is providing the translation of markdown source into react elements.
TypeScript is used to allow for catching type related bugs and automatic generation of typings files.
tsdx is mananging the build process.
storybook is leveraged to create runnable examples which can be hosted on GitHub pages.