Avoid bad hooks dependency usage in MarkdownHooks
As some comments point out in #890, there are some bad React hook usages
in those changes. A `useEffect()`, `useMemo()`, or `useCallback()`
should always specify all of its dependencies in the dependency array.
But the dependencies need to be referentially equal in order to not
trigger the hook, which is often not possible, given that plugin arrays
are often created during render.

This change replaces the `useEffect()` logic with a `useRef()`. Using
`useRef()`, we can persist an object instance across rerenders. We can
then perform our own props comparison logic to update or not instead of
relying on a dependency array. We also avoid `setState()`, which
triggered another render and could cause flashes of empty content.

The processor is updated if `remarkRehypeOptions`, `rehypePlugins`, or
`remarkPlugins` changed. The promise is updated if the processor or
`children` has changed.